refactor(web): extract ContactForm component (ARC-575)
ContactView was holding the form's state, action wiring, success-
card transition, and the SubmitButton helper inline alongside the
hero/stats/channels/side/faq composition. Pull all of that into a
dedicated ContactForm.tsx so the form owns its own state and the
page-level orchestrator stays at ~245 lines (down from ~455).

ContactForm exposes a single prop: the form i18n bundle. The
useActionState + useFormStatus wiring, the dismissal-by-identity
trick, and the SubmitButton wrapper now live in one file. No
behavior change.
