+++ apps/web/src/app/contact/ContactView.tsx
@@ -1,25 +1,14 @@
 'use client';
 
-import {
-  useActionState,
-  useState,
-  type CSSProperties,
-  type ReactNode,
-} from 'react';
-import { useFormStatus } from 'react-dom';
+import { type CSSProperties, type ReactNode } from 'react';
 import { useLanguage } from '@/shared/i18n/context';
 import { PageLayout } from '@arcadeum/ui/components/PageLayout/PageLayout';
 import { Container } from '@arcadeum/ui/components/Container/Container';
-import { GlassCard } from '@arcadeum/ui/components/GlassCard/GlassCard';
 import { Typography } from '@arcadeum/ui/components/Typography/Typography';
-import { Card } from '@arcadeum/ui/components/Card/Card';
 import { appConfig } from '@/shared/config/app-config';
 import { ActivityTicker } from '@arcadeum/ui/components/ActivityTicker';
 import { ChannelTile } from '@arcadeum/ui/components/ChannelTile';
 import { StatTile } from '@arcadeum/ui/components/StatTile';
-import { FloatingLabelInput } from '@arcadeum/ui/components/FloatingLabelInput';
-import { FloatingLabelTextArea } from '@arcadeum/ui/components/FloatingLabelTextArea';
-import { LaunchButton } from '@arcadeum/ui/components/LaunchButton';
 import { XStack, YStack } from 'tamagui';
 import { formatMessage } from '@/shared/i18n';
 import type { ContactMessages } from '@/shared/i18n/messages/legal/types';
@@ -35,7 +24,7 @@ import { useContactStyles } from './useContactStyles';
 import { ContactSidePanel } from './ContactSidePanel';
 import { ContactFaq, getFaqItems } from './ContactFaq';
 import { ContactAvatars } from './ContactAvatars';
-import { initialContactActionState, submitContactAction } from './actions';
+import { ContactForm } from './ContactForm';
 
 export interface ContactViewProps {
   t?: ContactMessages;
@@ -57,21 +46,6 @@ function HeroPill({
       {icon ? <span aria-hidden="true">{icon}</span> : null}
       {children}
     </span>
-  );
-}
-
-function SubmitButton({
-  idleLabel,
-  sendingLabel,
-}: {
-  idleLabel: string;
-  sendingLabel: string;
-}) {
-  const { pending } = useFormStatus();
-  return (
-    <LaunchButton isLaunching={pending} data-testid="contact-submit-button">
-      {pending ? sendingLabel : idleLabel}
-    </LaunchButton>
   );
 }
 
@@ -129,36 +103,6 @@ export default function ContactView({
   const side = sections?.side;
   const ticker = sections?.ticker;
   const faq = sections?.faq;
-
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
-  const [subject, setSubject] = useState('');
-  const [message, setMessage] = useState('');
-
-  const [actionState, formAction] = useActionState(
-    submitContactAction,
-    initialContactActionState,
-  );
-
-  // Track which action-state object the user has dismissed. New submissions
-  // produce a fresh state object, so identity comparison gives us a fresh
-  // success card per submit without setState-in-effect.
-  const [dismissedState, setDismissedState] = useState<
-    typeof actionState | null
-  >(null);
-
-  const fieldErrors =
-    actionState.status === 'invalid' ? actionState.fieldErrors : undefined;
-  const showSuccess =
-    actionState.status === 'ok' && actionState !== dismissedState;
-
-  const reset = () => {
-    setName('');
-    setEmail('');
-    setSubject('');
-    setMessage('');
-    setDismissedState(actionState);
-  };
 
   const titleParts = (hero?.title ?? '').trim().split(/\s+/);
   const lastWord = titleParts.pop() ?? '';
@@ -318,126 +262,8 @@ export default function ContactView({
             gap="$5"
           >
             <YStack flex={1.6} minWidth={0}>
-              <GlassCard>
-                <div style={s.formCardInnerStyle}>
-                  <div style={s.formHeaderStyle}>
-                    <YStack gap={2}>
-                      <span style={s.labelChipStyle}>
-                        {form?.subtitle ?? 'Direct message'}
-                      </span>
-                      <Typography variant="heading" uiSize="xl">
-                        {form?.title ?? 'Send the team a message'}
-                      </Typography>
-                    </YStack>
-                    <XStack alignItems="center" gap="$2">
-                      <ContactAvatars count={3} size={26} />
-                      <Typography variant="caption" alpha="medium">
-                        {form?.repliesNote ?? 'Replies hit your email'}
-                      </Typography>
-                    </XStack>
-                  </div>
-                  <hr style={s.ruleStyle} aria-hidden="true" />
-                  {showSuccess ? (
-                    <Card variant="glass" data-testid="contact-success-message">
-                      <div style={s.successCardStyle}>
-                        <div aria-hidden="true" style={s.burstStyle}>
-                          ✦
-                        </div>
-                        <Typography variant="heading" uiSize="lg">
-                          {form?.successTitle ??
-                            form?.success ??
-                            'Message away.'}
-                        </Typography>
-                        <Typography
-                          variant="body"
-                          alpha="medium"
-                          marginTop="$2"
-                        >
-                          {form?.successBody ??
-                            'Expect a reply within 4 hours. We sent a copy to your email.'}
-                        </Typography>
-                        <YStack alignItems="center" marginTop="$4">
-                          <button
-                            type="button"
-                            onClick={reset}
-                            style={s.helpLinkStyle}
-                          >
-                            {form?.sendAnother ?? 'Send another'}
-                          </button>
-                        </YStack>
-                      </div>
-                    </Card>
-                  ) : (
-                    <form action={formAction}>
-                      <YStack gap="$4">
-                        <div style={s.formGridStyle}>
-                          <FloatingLabelInput
-                            id="contact-name"
-                            name="name"
-                            label={form?.name ?? form?.nameLabel ?? 'Your name'}
-                            value={name}
-                            onChange={setName}
-                            required
-                            autoComplete="name"
-                            error={!!fieldErrors?.name}
-                            data-testid="contact-name-input"
-                          />
-                          <FloatingLabelInput
-                            id="contact-email"
-                            name="email"
-                            type="email"
-                            label={form?.email ?? form?.emailLabel ?? 'Email'}
-                            value={email}
-                            onChange={setEmail}
-                            required
-                            autoComplete="email"
-                            error={!!fieldErrors?.email}
-                            data-testid="contact-email-input"
-                          />
-                        </div>
-                        <FloatingLabelInput
-                          id="contact-subject"
-                          name="subject"
-                          label={
-                            form?.subject ?? form?.subjectLabel ?? 'Subject'
-                          }
-                          value={subject}
-                          onChange={setSubject}
-                          required
-                          error={!!fieldErrors?.subject}
-                          data-testid="contact-subject-input"
-                        />
-                        <FloatingLabelTextArea
-                          id="contact-message"
-                          name="message"
-                          label={
-                            form?.message ?? form?.messageLabel ?? 'Message'
-                          }
-                          value={message}
-                          onChange={setMessage}
-                          required
-                          maxLength={1200}
-                          error={!!fieldErrors?.message}
-                          data-testid="contact-message-textarea"
-                        />
-                        <div style={s.submitRowStyle}>
-                          <span style={s.privacyStyle}>
-                            <span aria-hidden="true">🔒</span>
-                            {form?.privacy ??
-                              'Private — we never share your email.'}
-                          </span>
-                          <SubmitButton
-                            idleLabel={form?.submit ?? 'Launch message'}
-                            sendingLabel={form?.submitting ?? 'Sending…'}
-                          />
-                        </div>
-                      </YStack>
-                    </form>
-                  )}
-                </div>
-              </GlassCard>
+              <ContactForm form={form} />
             </YStack>
-
             <ContactSidePanel side={side} workingHours={WORKING_HOURS} />
           </YStack>
